+++ docs/index.rst
@@ -32,7 +32,7 @@ Sections
 
    api/nucleus/index
    api/nucleus/modelci/index
-   api/nucleus/experimental/index
+   api/nucleus/deploy/index
 
 
 Index
